+++ src/main/java/com/Om/DentalClinic/service/PatientInfoService.java
@@ -10,7 +10,7 @@ public interface PatientInfoService {
 
 	public List<PatientInfo> getAllPatientInfo();
 
-	public void savePatientInfo(MultipartFile patientReports,String firstname,String middlename,String lastname,int patientage, String patientgender,Date patientregdate,int patientmobile1,int patientmobile2,String patientmedicalhistory) throws IOException;
+	public void savePatientInfo(MultipartFile patientReports,String firstname,String middlename,String lastname,int patientage, String patientgender,Date patientregdate,long patientmobile1,long patientmobile2,String patientmedicalhistory) throws IOException;
 	public PatientInfo getPatientInfoById(int id) ;
 	public void deletePatientInfoById(int id) ;
 //	public void savePatientInfo(PatientInfo patientInfo);
